refactor: make classes for different modes
