+++ docs/index.html
@@ -59,7 +59,7 @@
     </style>
 </head>
 
-<body>
+<body data-mode="default">
     <div class="container">
         <header class="page-header">
             <h1 class="title">Framework for ODRL Rule Compliance through Evaluation</h1>
