+++ docs/odrl3.0.html
@@ -61,7 +61,7 @@
     </style>
 </head>
 
-<body>
+<body data-mode="odrl3">
 
 
 
@@ -86,7 +86,7 @@
         </p>
         <div class="dropdown-container">
         <label for="dropdown">Select an ODRL 3.0 feature:</label>
-            <select id="odrl3-dropdown">
+            <select id="dropdown">
                 <!-- Options will be dynamically populated -->
             </select>
         </div>
