+++ src/util/prefixes.js
@@ -1,0 +1,10 @@
+export const prefixes = {
+    'odrl': 'http://www.w3.org/ns/odrl/2/',
+    'ex': 'http://example.org/',
+    'temp': 'http://example.com/request/',
+    'dct': 'http://purl.org/dc/terms/',
+    'xsd': 'http://www.w3.org/2001/XMLSchema#',
+    'foaf': 'http://xmlns.com/foaf/0.1/',
+    'report': 'https://w3id.org/force/compliance-report#',
+    'odrl3proposal': 'https://w3id.org/force/odrl3proposal#',
+};
